requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
datasets>=2.14.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
newspaper3k>=0.2.8
tqdm>=4.65.0

# App
streamlit>=1.28.0

# Utilities
joblib>=1.3.0
